+++ app-partial-users/src/main/java/com/jraska/github/client/users/UsersAppComponent.kt
@@ -5,9 +5,8 @@ import android.content.Context
 import androidx.lifecycle.ViewModelProvider
 import com.facebook.drawee.backends.pipeline.Fresco
 import com.jakewharton.threetenabp.AndroidThreeTen
-import com.jraska.github.client.FakeConfigModule
 import com.jraska.github.client.FakeCoreModule
-import com.jraska.github.client.WebLinkLauncher
+import com.jraska.github.client.FakeWebLinkModule
 import com.jraska.github.client.core.android.CoreAndroidModule
 import com.jraska.github.client.core.android.OnAppCreate
 import com.jraska.github.client.http.HttpModule
@@ -16,7 +15,6 @@ import dagger.Component
 import dagger.Module
 import dagger.Provides
 import dagger.multibindings.IntoSet
-import okhttp3.HttpUrl
 import javax.inject.Singleton
 
 @Singleton
@@ -25,8 +23,8 @@ import javax.inject.Singleton
     AppModule::class,
     CoreAndroidModule::class,
     FakeCoreModule::class,
-    FakeConfigModule::class,
     HttpModule::class,
+    FakeWebLinkModule::class,
     UsersModule::class]
 )
 interface UsersAppComponent {
@@ -42,15 +40,6 @@ interface UsersAppComponent {
 
 @Module
 object AppModule {
-  @Provides
-  fun provideLinkLauncher(): WebLinkLauncher {
-    return object : WebLinkLauncher {
-      override fun launchOnWeb(url: HttpUrl) {
-        throw UnsupportedOperationException("Partial app tried to open $url")
-      }
-    }
-  }
-
   @Provides
   @IntoSet
   fun setupFresco(): OnAppCreate {
