+++ app/src/main/java/com/jraska/github/client/AppComponent.kt
@@ -54,8 +54,7 @@ interface AppComponent {
     PushModule::class,
     SettingsModule::class,
     AboutModule::class,
-    ShortcutsModule::class,
-    FirebaseConfigModule::class]
+    ShortcutsModule::class]
 )
 object SharedModules
