+++ app/src/main/java/com/jraska/github/client/FirebaseCoreModule.kt
@@ -45,4 +45,19 @@ object FirebaseCoreModule {
   internal fun firebaseCrash(): CrashReporter {
     return FirebaseCrashlyticsReporter()
   }
+
+  @Provides
+  @Singleton
+  internal fun config(decorations: Set<@JvmSuppressWildcards Config.Decoration>): Config {
+    val configProxy = FirebaseConfigProxy(FirebaseRemoteConfig.getInstance())
+
+    configProxy.setupDefaults().fetch()
+
+    var config: Config = configProxy
+    decorations.forEach {
+      config = it.decorate(config)
+    }
+
+    return config
+  }
 }
