--- core-testing/src/main/java/com/jraska/github/client/FakeConfigModule.kt
@@ -1,10 +1,0 @@
-package com.jraska.github.client
-
-import dagger.Module
-import dagger.Provides
-
-@Module
-object FakeConfigModule {
-  @Provides
-  fun config(): Config = FakeCoreModule.config
-}
