+++ core-testing/src/main/java/com/jraska/github/client/FakeCoreModule.kt
@@ -6,16 +6,30 @@ import com.jraska.github.client.logging.CrashReporter
 import dagger.Module
 import dagger.Provides
 import dagger.multibindings.IntoSet
+import javax.inject.Singleton
 
 @Module
 object FakeCoreModule {
   val eventAnalytics = Fakes.recordingAnalytics()
-  val config = Fakes.config()
   val analyticProperty = Fakes.recordingAnalyticsProperty()
 
   @Provides
   fun analytics(): EventAnalytics {
     return eventAnalytics
+  }
+
+  @Provides
+  @Singleton
+  fun fakeConfig(): FakeConfig = Fakes.config()
+
+  @Provides
+  fun config(fakeConfig: FakeConfig, decorations: Set<@JvmSuppressWildcards Config.Decoration>): Config {
+    var config: Config = fakeConfig
+    decorations.forEach {
+      config = it.decorate(config)
+    }
+
+    return config
   }
 
   @Provides
@@ -30,10 +44,10 @@ object FakeCoreModule {
 
   @Provides
   @IntoSet
-  fun config(): Config.Decoration {
+  fun configDecoration(): Config.Decoration {
     return object : Config.Decoration {
       override fun decorate(originalConfig: Config): Config {
-        return config
+        return originalConfig
       }
     }
   }
