+++ core-testing/src/main/java/com/jraska/github/client/FakeWebLinkModule.kt
@@ -1,0 +1,17 @@
+package com.jraska.github.client
+
+import dagger.Module
+import dagger.Provides
+import javax.inject.Singleton
+
+@Module
+object FakeWebLinkModule {
+  @Provides
+  @Singleton
+  fun provideRecordingLinkLauncher(): RecordingWebLinkLauncher {
+    return Fakes.recordingWebLinkLauncher()
+  }
+
+  @Provides
+  fun provideLinkLauncher(recordingLauncher: RecordingWebLinkLauncher): WebLinkLauncher = recordingLauncher
+}
