+++ feature/repo/src/test/java/com/jraska/github/client/repo/di/TestRepoComponent.kt
@@ -1,6 +1,6 @@
 package com.jraska.github.client.repo.di
 
-import com.jraska.github.client.FakeModules
+import com.jraska.github.client.FakeUnitTestModules
 import com.jraska.github.client.android.FakeSnackbarDisplay
 import com.jraska.github.client.repo.RepoDetailViewModel
 import com.jraska.github.client.repo.RepoModule
@@ -9,7 +9,7 @@ import okhttp3.mockwebserver.MockWebServer
 import javax.inject.Singleton
 
 @Singleton
-@Component(modules = [RepoModule::class, FakeModules::class])
+@Component(modules = [RepoModule::class, FakeUnitTestModules::class])
 internal interface TestRepoComponent {
   fun repoDetailViewModel(): RepoDetailViewModel
