+++ feature/users/src/test/java/com/jraska/github/client/users/di/TestUsersComponent.kt
@@ -1,6 +1,6 @@
 package com.jraska.github.client.users.di
 
-import com.jraska.github.client.FakeModules
+import com.jraska.github.client.FakeUnitTestModules
 import com.jraska.github.client.users.UserDetailViewModel
 import com.jraska.github.client.users.UsersModule
 import com.jraska.github.client.users.UsersViewModel
@@ -10,7 +10,7 @@ import javax.inject.Singleton
 
 @Singleton
 @Component(
-  modules = [FakeModules::class, UsersModule::class]
+  modules = [FakeUnitTestModules::class, UsersModule::class]
 )
 internal interface TestUsersComponent {
   fun usersViewModel(): UsersViewModel
